// src/lib/types.ts
export interface Tutor {
    id: string;
    name: string;
    subjects: string[];
    experience: number;
    location: string;
    timings: string;
    priceRange: number;
    languages: string[];
    rating: number;
    demoVideos: string[];
    notes: string[];
    bio: string;
    image: string;
}